scuwdt: Fixed MISRAC rule 10.4 violations
This patch fix following violation
"misra_c_2012_rule_10_4_violation: Essential type of the left hand
operand Index is not the same as that of the right operand."


Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/scuwdt/src/xscuwdt_sinit.c
@@ -54,7 +54,7 @@ XScuWdt_Config *XScuWdt_LookupConfig(u16 DeviceId)
 	XScuWdt_Config *CfgPtr = NULL;
 	u32 Index;
 
-	for (Index = 0U; Index < XPAR_XSCUWDT_NUM_INSTANCES; Index++) {
+	for (Index = 0U; Index < (u32)XPAR_XSCUWDT_NUM_INSTANCES; Index++) {
 		if (XScuWdt_ConfigTable[Index].DeviceId == DeviceId) {
 			CfgPtr = &XScuWdt_ConfigTable[Index];
 			break;
